refactor: 수색 선택 컴포넌트 리팩토링
